# Reuse AggregatorHandle with cumulative temporality