--- a/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/AggregatorHandle.java
+++ b/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/AggregatorHandle.java
@@ -85,22 +85,22 @@ public abstract class AggregatorHandle<T, U extends ExemplarData> implements Bou
   }
 
   /**
-   * Returns the current value into as {@link T} and resets the current value in this {@code
-   * Aggregator}.
+   * Returns the current value into as {@link T}. If {@code reset} is {@code true}, resets the
+   * current value in this {@code Aggregator}.
    */
   @Nullable
-  public final T accumulateThenReset(Attributes attributes, boolean reset) {
+  public final T accumulateThenMaybeReset(Attributes attributes, boolean reset) {
     if (!hasRecordings) {
       return null;
     }
     if (reset) {
       hasRecordings = false;
     }
-    return doAccumulateThenReset(exemplarReservoir.collectAndReset(attributes), reset);
+    return doAccumulateThenMaybeReset(exemplarReservoir.collectAndReset(attributes), reset);
   }
 
-  /** Implementation of the {@code accumulateThenReset}. */
-  protected abstract T doAccumulateThenReset(List<U> exemplars, boolean reset);
+  /** Implementation of the {@link #accumulateThenMaybeReset(Attributes, boolean)}. */
+  protected abstract T doAccumulateThenMaybeReset(List<U> exemplars, boolean reset);
 
   @Override
   public final void recordLong(long value, Attributes attributes, Context context) {