--- a/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/DoubleExplicitBucketHistogramAggregator.java
+++ b/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/DoubleExplicitBucketHistogramAggregator.java
@@ -122,7 +122,7 @@ public final class DoubleExplicitBucketHistogramAggregator
     }
 
     @Override
-    protected ExplicitBucketHistogramAccumulation doAccumulateThenReset(
+    protected ExplicitBucketHistogramAccumulation doAccumulateThenMaybeReset(
         List<DoubleExemplarData> exemplars, boolean reset) {
       lock.lock();
       try {