--- a/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/DropAggregator.java
+++ b/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/DropAggregator.java
@@ -37,7 +37,8 @@ public final class DropAggregator implements Aggregator<Object, DoubleExemplarDa
         protected void doRecordDouble(double value) {}
 
         @Override
-        protected Object doAccumulateThenReset(List<DoubleExemplarData> exemplars, boolean reset) {
+        protected Object doAccumulateThenMaybeReset(
+            List<DoubleExemplarData> exemplars, boolean reset) {
           return ACCUMULATION;
         }
       };