--- a/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/LongSumAggregator.java
+++ b/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/LongSumAggregator.java
@@ -94,8 +94,12 @@ public final class LongSumAggregator
     }
 
     @Override
-    protected LongAccumulation doAccumulateThenReset(List<LongExemplarData> exemplars) {
-      return LongAccumulation.create(this.current.sumThenReset(), exemplars);
+    protected LongAccumulation doAccumulateThenReset(
+        List<LongExemplarData> exemplars, boolean reset) {
+      if (reset) {
+        return LongAccumulation.create(this.current.sumThenReset(), exemplars);
+      }
+      return LongAccumulation.create(this.current.sum(), exemplars);
     }
 
     @Override