--- a/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/LongSumAggregator.java
+++ b/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/LongSumAggregator.java
@@ -46,13 +46,6 @@ public final class LongSumAggregator
   }
 
   @Override
-  public LongAccumulation merge(
-      LongAccumulation previousAccumulation, LongAccumulation accumulation) {
-    return LongAccumulation.create(
-        previousAccumulation.getValue() + accumulation.getValue(), accumulation.getExemplars());
-  }
-
-  @Override
   public LongAccumulation diff(
       LongAccumulation previousAccumulation, LongAccumulation accumulation) {
     return LongAccumulation.create(