--- a/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/state/DefaultSynchronousMetricStorage.java
+++ b/sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/state/DefaultSynchronousMetricStorage.java
@@ -5,8 +5,6 @@
 
 package io.opentelemetry.sdk.metrics.internal.state;
 
-import static io.opentelemetry.sdk.metrics.internal.state.MetricStorageUtils.MAX_ACCUMULATIONS;
-
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
@@ -16,6 +14,7 @@ import io.opentelemetry.sdk.metrics.data.ExemplarData;
 import io.opentelemetry.sdk.metrics.data.MetricData;
 import io.opentelemetry.sdk.metrics.internal.aggregator.Aggregator;
 import io.opentelemetry.sdk.metrics.internal.aggregator.AggregatorHandle;
+import io.opentelemetry.sdk.metrics.internal.aggregator.EmptyMetricData;
 import io.opentelemetry.sdk.metrics.internal.descriptor.MetricDescriptor;
 import io.opentelemetry.sdk.metrics.internal.export.RegisteredReader;
 import io.opentelemetry.sdk.metrics.internal.view.AttributesProcessor;
@@ -42,10 +41,10 @@ public final class DefaultSynchronousMetricStorage<T, U extends ExemplarData>
 
   private final RegisteredReader registeredReader;
   private final MetricDescriptor metricDescriptor;
+  private final AggregationTemporality aggregationTemporality;
   private final Aggregator<T, U> aggregator;
   private final ConcurrentHashMap<Attributes, AggregatorHandle<T, U>> activeCollectionStorage =
       new ConcurrentHashMap<>();
-  private final TemporalMetricStorage<T, U> temporalMetricStorage;
   private final AttributesProcessor attributesProcessor;
 
   DefaultSynchronousMetricStorage(
@@ -55,18 +54,11 @@ public final class DefaultSynchronousMetricStorage<T, U extends ExemplarData>
       AttributesProcessor attributesProcessor) {
     this.registeredReader = registeredReader;
     this.metricDescriptor = metricDescriptor;
-    AggregationTemporality aggregationTemporality =
+    this.aggregationTemporality =
         registeredReader
             .getReader()
             .getAggregationTemporality(metricDescriptor.getSourceInstrument().getType());
     this.aggregator = aggregator;
-    this.temporalMetricStorage =
-        new TemporalMetricStorage<>(
-            aggregator,
-            /* isSynchronous= */ true,
-            registeredReader,
-            aggregationTemporality,
-            metricDescriptor);
     this.attributesProcessor = attributesProcessor;
   }
 
@@ -126,6 +118,7 @@ public final class DefaultSynchronousMetricStorage<T, U extends ExemplarData>
         }
         // Try to remove the boundAggregator. This will race with the collect method, but only one
         // will succeed.
+        // TODO: is this still valid?
         activeCollectionStorage.remove(attributes, boundAggregatorHandle);
         continue;
       }
@@ -165,24 +158,39 @@ public final class DefaultSynchronousMetricStorage<T, U extends ExemplarData>
       InstrumentationScopeInfo instrumentationScopeInfo,
       long startEpochNanos,
       long epochNanos) {
+    boolean reset = aggregationTemporality == AggregationTemporality.DELTA;
+
     // Grab accumulated measurements.
     Map<Attributes, T> accumulations = new HashMap<>();
     for (Map.Entry<Attributes, AggregatorHandle<T, U>> entry : activeCollectionStorage.entrySet()) {
-      boolean unmappedEntry = entry.getValue().tryUnmap();
-      if (unmappedEntry) {
-        // If able to unmap then remove the record from the current Map. This can race with the
-        // acquire but because we requested a specific value only one will succeed.
-        activeCollectionStorage.remove(entry.getKey(), entry.getValue());
+      if (reset) {
+        boolean unmappedEntry = entry.getValue().tryUnmap();
+        if (unmappedEntry) {
+          // If able to unmap then remove the record from the current Map. This can race with the
+          // acquire but because we requested a specific value only one will succeed.
+          activeCollectionStorage.remove(entry.getKey(), entry.getValue());
+        }
       }
-      T accumulation = entry.getValue().accumulateThenReset(entry.getKey());
+      T accumulation = entry.getValue().accumulateThenReset(entry.getKey(), reset);
       if (accumulation == null) {
         continue;
       }
       accumulations.put(entry.getKey(), accumulation);
     }
 
-    return temporalMetricStorage.buildMetricFor(
-        resource, instrumentationScopeInfo, accumulations, startEpochNanos, epochNanos);
+    if (accumulations.isEmpty()) {
+      return EmptyMetricData.getInstance();
+    }
+
+    return aggregator.toMetricData(
+        resource,
+        instrumentationScopeInfo,
+        metricDescriptor,
+        accumulations,
+        aggregationTemporality,
+        startEpochNanos,
+        registeredReader.getLastCollectEpochNanos(),
+        epochNanos);
   }
 
   @Override